# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Service API
  description: API for managing services
  version: 1.0.0
servers:
  - url: http://localhost:8000
paths:
  /service:
    post:
      summary: Add a new service
      description: Add a new service with name, state, and description
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Service'
      responses:
        '201':
          description: Service created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Service'
  /services:
    get:
      summary: Get all services
      description: Retrieve a list of all services
      responses:
        '200':
          description: List of services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Service'
  /service/{name}:
    get:
      summary: Get service history
      description: Get the history of a specific service by name
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Service history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Service'
  /sla/{name}:
    get:
      summary: Get SLA for a service
      description: Calculate the SLA for a service over a given interval
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
        - name: interval
          in: query
          required: true
          schema:
            type: string
            example: '24h'
      responses:
        '200':
          description: SLA value
          content:
            application/json:
              schema:
                type: object
                properties:
                  sla:
                    type: number
components:
  schemas:
    Service:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        state:
          type: string
        description:
          type: string
        timestamp:
          type: string
          format: date-time
        timestamp_end:
          type: string
          format: date-time